# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/departments/{id}": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Get a Department",
                "description": "returns the details of a single Department",
                "operationId": "8d8e7a213e3fdd9c2ab7a3297ddef213",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the department",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unit found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/department_units/{id}": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Get a Department",
                "description": "Returns the details of a single Department",
                "operationId": "326c7b38924df5ef223546e90ae7a08e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the department",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unit found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DepartmentUnitResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/processflowhistory": {
            "get": {
                "tags": [
                    "processflow History"
                ],
                "summary": "Fetch all processflow histories",
                "operationId": "0ca1cf3ff71d62d5c4277c223dbb330b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProcessFlowHistoryResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/processflowhistory/create": {
            "post": {
                "tags": [
                    "processflow History"
                ],
                "summary": "Create a new processflow history",
                "operationId": "b1221484d1cbd8eecfc45fd8c326204b",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/processflowhistory/{id}": {
            "get": {
                "tags": [
                    "processflow History"
                ],
                "summary": "Fetch a processflow history",
                "operationId": "d2926d013ac09d2010aec24a9ec5ad2a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProcessFlowHistoryResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "processflow History"
                ],
                "summary": "Delete a processflow history",
                "description": "Deletes a processflowhistory by its ID.",
                "operationId": "deleteprocessflowHistory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the route to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Route not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DepartmentUnitResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "unit": {
                        "$ref": "#/components/schemas/UnitResource"
                    }
                },
                "type": "object"
            },
...
            "ProcessFlowHistoryResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "task_id": {
                        "type": "integer"
                    },
                    "step_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "process_flow_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "UnitResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        }
    },
...
}